physics model set up. happy now?

+++ types.h
@@ -44,7 +44,7 @@ typedef struct layer
 	s32 x,y;
 	u16 w,h;
 	tile_t *tiles;
-	u8 *updmask;
+	u32 *updmask;
 } layer_t;
 
 typedef struct netpacket netpacket_t;
